requirements/base.txt
# Include the requirements we're caching as "wheel" archives
# to speed up the test builds.
-r wheels.txt

# edX Internal Requirements
git+https://github.com/edx/XBlock.git@xblock-0.4.12#egg=XBlock==0.4.12

# edx-submissions
git+https://github.com/edx/edx-submissions.git@1.1.2#egg=edx-submissions==1.1.2

# Third Party Requirements
boto>=2.32.1,<3.0.0
python-swiftclient==3.1.0
celery==3.1.18
defusedxml==0.4.1
django>=1.8.14,<1.9
django-extensions==1.5.9
django-model-utils==2.3.1
djangorestframework>=3.1,<3.3
dogapi==1.2.1
jsonfield==1.0.3
lazy==1.1
libsass==0.10.0
loremipsum==1.0.5
python-dateutil==2.1
python-memcached==1.48
pytz==2015.2
voluptuous==0.8.5

# AI grading
git+https://github.com/edx/ease.git@release-2015-07-14#egg=ease==0.1.3
git+https://github.com/edx/nltk.git@2.0.6#egg=nltk==2.0.6
